Split georef headers
